Consume discount grpc server as grpc client in Basket.API

+++ src/Services/Basket/Basket.API/Models/ShoppingCart.cs
@@ -14,5 +14,5 @@ public class ShoppingCart
 
     public string UserName { get; set; } = default!;
     public List<ShoppingCartItem> Items { get; set; } = new();
-    public decimal TotalPrice => Items.Sum(i => i.Price * i.Quantity);
+    public double TotalPrice => Items.Sum(i => i.Price * i.Quantity);
 }
